source/backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
authx==1.5.1
bcrypt==5.0.0
boto3==1.42.75
botocore==1.42.75
cffi==2.0.0
click==8.3.1
cryptography==46.0.5
ecdsa==0.19.1
fastapi==0.135.1
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
jmespath==1.1.0
psycopg==3.3.3
psycopg-binary==3.3.3
pyasn1==0.6.3
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.22
pytz==2025.2
rsa==4.9.1
s3transfer==0.16.0
six==1.17.0
SQLAlchemy==2.0.48
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
python-docx==1.2.0
openpyxl==3.1.5
